Added comments, extracted css and fixed email validation bug
Added comments, extracted css and fixed email validation bug
