+++ src/features/home/HomePage.js
@@ -1,7 +1,9 @@
 import React, { useEffect, useState } from "react";
-import agent from "../api/agent/agent";
+import agent from "../../app/api/agent/agent";
 import { useNavigate } from 'react-router-dom';
-import Loading from "../layout/Loading";
+import Loading from "../../app/layout/Loading";
+import { FaHistory } from "react-icons/fa";
+import FixedButton from "../../app/common/FixedButton";
 
 function HomePage() {
     const [loading, setLoading] = useState(false);
@@ -25,15 +27,14 @@ function HomePage() {
         checkServerHealth();
     }, []);
 
-    return (
-        <>
-            { loading ? <Loading/> : 
-                <button onClick={checkServerHealth}>
-                    Retry Action
-                </button>
-            }
-        </>
-    );
+    return <>{ loading ? <Loading/> : 
+                <FixedButton 
+                    onClick={checkServerHealth}
+                    iconComponent={FaHistory}
+                    callback={checkServerHealth}
+                    label={"Retry"}
+                />                
+            }</>
 }
 
 export default HomePage;
