+++ src/features/question/QuestionDetailsPage.js
@@ -1,14 +1,15 @@
 import React, {useEffect, useState} from "react";
 import { observer } from "mobx-react-lite";
-import QuestionDetails from "../../features/question/QuestionDetails";
+import QuestionDetails from "./QuestionDetails";
 import { FaArrowCircleLeft } from 'react-icons/fa';
 import { useParams, useNavigate } from "react-router-dom";
-import Loading from "../layout/Loading";
-import FixedButton from "../common/FixedButton";
+import Loading from "../../app/layout/Loading";
+import FixedButton from "../../app/common/FixedButton";
 import { useStore } from "../stores/store";
 
 
 export default observer(function QuestionDetailsPage() {
+    //Using MOBX question store to remove alot of the logic from the components
     const {questionStore} = useStore();
     const {loadQuestion , loadingDetails, question, submitting, message, editQuestion} = questionStore;
     const [selectedChoice, setChoice] = useState('');
