+++ src/features/question/QuestionListItem.js
@@ -1,9 +1,9 @@
 import {useContext} from "react";
 import Panel from "../../app/common/Panel";
 import { useNavigate } from "react-router-dom";
-import { ModalContext } from "../../app/stores/ModalStore";
+import { ModalContext } from "../stores/ModalStore";
 
-function QuestionListItem({question}) {
+function QuestionListItem({question, index}) {
     const navigate = useNavigate();
     const [state, setState] = useContext(ModalContext);
 
@@ -15,24 +15,17 @@ function QuestionListItem({question}) {
     const contentUrl = window.location.href + '/' + question.id.toString();
 
     return (
-        <Panel key={"question-" + question.id} className="m-8">
+        <Panel key={"question-" + index} className="m-8">
             <div className="list-item">
-                <img 
-                    className="list-item-thumb" 
-                    src={question.thumb_url}
-                />
+                <img className="list-item-thumb" src={question.thumb_url}/>
                 <div className="list-item-content">
-                    <a 
-                        className="list-item-header"
-                        onClick={() => navigate(`/questions/${question.id}`)}
+                    <a className="list-item-header" onClick={() => navigate(`/questions/${question.id}`)}
                     >
                         {question.question}
                     </a>
                     <div className="list-item-date">Published at: {date ? date.split('+')[0] : 'N/A'}</div>
                 </div>
-                <button 
-                    onClick={() => onModalVisibleChange(contentUrl)}
-                >
+                <button onClick={() => onModalVisibleChange(contentUrl)}>
                     Share
                 </button>
             </div>
