--- src/index.css
@@ -1,466 +1,0 @@
-body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
-
-/*common*/
-.m-8 {
-  margin: 8px !important;
-}
-
-/*Loading Screen */
-.loader-container {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: fixed;
-  background: #c9c5e2;
-  z-index: 1;
-  opacity: 0.8;
-}
-
-.loader {
-  width: 82px;
-  height: 82px;
-  border: 10px solid #FFF;
-  border-bottom-color: #2a58e2;
-  border-radius: 50%;
-  display: inline-block;
-  box-sizing: border-box;
-  animation: rotation 1.8s linear infinite;
-  }
-
-  @keyframes rotation {
-  0% {
-    transform: rotate(0deg);
-  }
-  50% {
-    border-bottom-color: #0bb00e;
-    transform: rotate(180deg);
-  }
-  100% {
-    border-bottom-color: #e2230e;
-    transform: rotate(360deg);
-  }
-} 
-
-
-/*Question List Item*/
-.list-item {
-  display: contents;
-  table-layout: fixed;
-  padding: 0.21428571em 0;
-  line-height: 1.14285714em;
-}
-
-.list-item button {
-  float: right;
-  color: white;
-  background-color: #2c6baa;
-  border: 1px solid #2c6baa;
-  border-radius: 4px;
-}
-
-.list-item-thumb {
-  display: inline-block;
-  background-color: transparent;
-  vertical-align: top;
-  width: 5rem;
-  height: 5rem;
-  border-radius: 500rem;
-}
-
-
-.list-item-header {
-  cursor: pointer;
-  color: #2c6baa!important;
-  display: block;
-  margin: 0;
-  font-family: Lato,'Helvetica Neue',Arial,Helvetica,sans-serif;
-  font-weight: 700;
-  color: rgba(0,0,0,.87);
-}
-
-.list-item-date {
-  display: block;
-  color: rgba(0,0,0,.7);
-  line-height: 1.14285714em;
-  font-size: 10px;
-}
-
-
-/* TEST CSS */
-
-.panel {
-  background: #fff;
-  box-shadow: 0 1px 2px 0 #2c6baa;
-  margin: 0 0;
-  padding: 8px;
-  border-radius: 0.4rem;
-  border: 1.5px solid rgba(34,36,38,.15)
-}
-
-.details-panel {
-  background: rgb(241, 244, 250);
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: stretch;
-  width: 100%;
-  max-width: 800px;
-  margin: 8px auto 8px auto;
-  padding: 20px;
-  box-sizing: border-box;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  box-shadow: 0 1px 2px 0 #2c6baa;
-}
-
-.mock-container button {
-  background-color: #2c6baa;
-  position: fixed;
-  top: 20px;
-  left: 20px;
-}
-
-.list-item-content {
-  display: inline-block;
-  width: auto;
-  padding: 10px 0 0 0.75rem;
-  vertical-align: top;
-  line-height: 1.14285714em;
-}
-
-.details-panel-header {
-  height: 15%;
-  width: auto;
-  box-sizing: border-box;
-  background: inherit;
-  align-items: center;
-  margin: 0px !important;
-  justify-items: center;
-  display: inline-grid;
-}
-
-.details-panel-header img {
-  display: block;
-  max-width: 100%;
-  height: auto;
-  border-radius: 4px;
-}
-
-.details-panel-body {
-  flex: 2;
-  margin-top: 20px;
-}
-
-
-.details-panel-body button {
-    position:relative;
-    float: right;
-    color: white;
-    background-color: #2c6baa;
-    border: 1px solid #2c6baa;
-    border-radius: 4px;
-    margin: 0 4px;  
-}
-
-.details-panel-body h2 {
-  margin-top: 0;
-  margin-bottom: 10px;
-  font-size: 24px;
-}
-
-.details-panel-body p {
-  margin: 0;
-  font-size: 18px;
-  line-height: 1.5;
-}
-
-.details-panel-poll {
-  margin-top: 20px;
-}
-
-.details-panel-poll h3 {
-  margin-top: 0;
-  margin-bottom: 10px;
-  font-size: 18px;
-}
-
-.details-panel-poll form label {
-  display: block;
-  margin-bottom: 10px;
-}
-
-.details-panel-footer {
-  margin-top: 20px;
-}
-
-.textbox {
-  background: #e5edf7
-}
-
-/*divider*/
-.divider-xs {
-  height: 1px;
-  background-color: #acd2f7;
-  margin: 0 4%;
-}
-
-.divider {
-  height: 1px;
-  background-color: #2c6baa;
-  margin: 8px 0;
-}
-
-
-@media (max-width: 768px) {
-  .divider {
-    margin: 4px 0;
-  }
-}
-
-/* Filter */
-
-.filter-input {
-  padding: 8px;
-  width: 100%;
-  justify-content: center;
-  display: flex !important;
-}
-
-.filter-input button {
-  position:relative;
-  float: right;
-  color: white;
-  background-color: #2c6baa;
-  border: 1px solid #2c6baa;
-  border-radius: 4px;
-  margin: 0 4px;  
-}
-
-.filter-label {
-  font-size: 24px;
-  font-weight: 600;
-  margin-right: 12px;
-  padding: 4px;
-}
-
-.loader-container-xs {
-  display: flex;
-  justify-content: center;
-  background-color: inherit;
-  margin: 8px;
-  padding: 8px;
-}
-
-
-
-
-/*
-MODAL
-*/
-
-.modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.modal {
-  background-color: white;
-  border-radius: 5px;
-  padding: 20px;
-  position: relative;
-  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
-  width: 400px;
-  max-width: 90%;
-  max-height: 90%;
-  overflow: auto;
-}
-
-.modal-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 20px;
-}
-
-.modal-body {
-  margin-bottom: 20px;
-}
-
-label {
-  display: block;
-  margin-bottom: 10px;
-}
-
-input {
-  display: block;
-  width: 80%;
-  padding: 10px;
-  border-radius: 5px;
-  border: 1px solid #ccc;
-}
-
-.modal-footer {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.cancel-button {
-  background-color: #adb0b8;
-  color: rgb(0, 0, 0);
-  border: none;
-  padding: 10px 20px;
-  border-radius: 4px;
-  cursor: pointer;
-  margin: 0px 5px;
-}
-
-.submit-button {
-  background-color: rgb(64, 64, 201);
-  color: white;
-  border: none;
-  padding: 10px 20px;
-  border-radius: 4px;
-  cursor: pointer;
-  margin: 0px 5px;
-}
-
-.red {
-  background-color: red !important
-}
-
-.green {
-  background-color: rgb(102, 189, 102) !important
-}
-
-.blue {
-  color: rgb(45, 45, 254) !important
-}
-
-button {
-  border-radius: 4px;
-}
-
-/* VOTING POLE */
-.voting-poll {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-  border: 2px solid #ccc;
-  border-radius: 5px;
-  margin-bottom: 20px;
-}
-
-.choices-container {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 20px;
-}
-
-.choice {
-  display: flex;
-  align-items: center;
-  margin-bottom: 10px;
-}
-
-label {
-  font-size: 16px;
-}
-
-.vote-button {
-  background-color: #00ff51;
-  color: #fff;
-  border: none;
-  padding: 10px 20px;
-  border-radius: 5px;
-  cursor: pointer;
-  margin-bottom: 20px;
-  transition: all 0.2s ease-in-out;
-}
-
-.vote-button:hover {
-  background-color: #0069d9;
-}
-
-.results-container {
-  display: flex;
-  flex-direction: column;
-}
-
-.result {
-  display: flex;
-  align-items: center;
-  margin-bottom: 10px;
-}
-
-.result-label {
-  font-size: 16px;
-  margin-right: 10px;
-}
-
-.result-bar {
-  background-color: #007bff;
-  color: #fff;
-  height: 30px;
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  padding-right: 10px;
-  border-top-right-radius: 5px;
-  border-bottom-right-radius: 5px;
-  transition: all 0.2s ease-in-out;
-}
-
-.result-bar:hover {
-  background-color: #0069d9;
-}
-
-.voting-poll-item-container {
-  background: #d0e1f2;
-  margin: 6px 0;
-  padding: 4px;
-  border-radius: 4px;
-  box-shadow: #2a58e2;
-  border: 1px solid black;
-}
-
-.voting-poll-item {
-  display: flex;
-}
-
-.voting-poll-item label {
-  width: 95%;
-  font-weight: bold;
-}
-
-.fixed-button {
-  position: fixed;
-  top: 25px;
-  left: 25px;
-  color: #2c6baa;
-}
-
-.hide {
-  display: none;
-}
