+++ src/index.js
@@ -1,18 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './styles/index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
-import { ModalProvider } from './app/stores/ModalStore';
+import { ModalProvider } from './features/stores/ModalStore';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
+
 root.render(
     <ModalProvider>
       <App />
     </ModalProvider>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
