+++ src/styles/filter.css
@@ -1,0 +1,23 @@
+.filter-input {
+    padding: 8px;
+    width: 100%;
+    justify-content: center;
+    display: flex !important;
+  }
+  
+  .filter-input button {
+    position:relative;
+    float: right;
+    color: white;
+    background-color: #2c6baa;
+    border: 1px solid #2c6baa;
+    border-radius: 4px;
+    margin: 0 4px;  
+  }
+  
+  .filter-label {
+    font-size: 24px;
+    font-weight: 600;
+    margin-right: 12px;
+    padding: 4px;
+  }
