+++ src/styles/index.css
@@ -1,0 +1,26 @@
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
+}
+
+/*common*/
+.m-8 {
+  margin: 8px !important;
+}
+
+@import './filter.css';
+@import './list.css';
+@import './modal.css';
+@import './votingPoll.css';
+@import './panel.css';
+@import './loader.css';
+@import './App.css';
