+++ src/styles/loader.css
@@ -1,0 +1,38 @@
+
+/*Loading Screen */
+.loader-container {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    position: fixed;
+    background: #c9c5e2;
+    z-index: 1;
+    opacity: 0.8;
+  }
+  
+  .loader {
+    width: 82px;
+    height: 82px;
+    border: 10px solid #FFF;
+    border-bottom-color: #2a58e2;
+    border-radius: 50%;
+    display: inline-block;
+    box-sizing: border-box;
+    animation: rotation 1.8s linear infinite;
+    }
+  
+    @keyframes rotation {
+    0% {
+      transform: rotate(0deg);
+    }
+    50% {
+      border-bottom-color: #0bb00e;
+      transform: rotate(180deg);
+    }
+    100% {
+      border-bottom-color: #e2230e;
+      transform: rotate(360deg);
+    }
+  }
