+++ src/styles/panel.css
@@ -1,0 +1,114 @@
+.panel {
+    background: #fff;
+    box-shadow: 0 1px 2px 0 #2c6baa;
+    margin: 0 0;
+    padding: 8px;
+    border-radius: 0.4rem;
+    border: 1.5px solid rgba(34,36,38,.15)
+}
+  
+.details-panel {
+    background: rgb(241, 244, 250);
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: stretch;
+    width: 100%;
+    max-width: 800px;
+    margin: 8px auto 8px auto;
+    padding: 20px;
+    box-sizing: border-box;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    box-shadow: 0 1px 2px 0 #2c6baa;
+  }
+  
+.mock-container button {
+    background-color: #2c6baa;
+    position: fixed;
+    top: 20px;
+    left: 20px;
+}
+
+  
+.details-panel-header {
+    height: 15%;
+    width: auto;
+    box-sizing: border-box;
+    background: inherit;
+    align-items: center;
+    margin: 0px !important;
+    justify-items: center;
+    display: inline-grid;
+}
+  
+.details-panel-header img {
+    display: block;
+    max-width: 100%;
+    height: auto;
+    border-radius: 4px;
+}
+  
+.details-panel-body {
+    flex: 2;
+    margin-top: 20px;
+}
+    
+.details-panel-body button {
+    position:relative;
+    float: right;
+    color: white;
+    background-color: #2c6baa;
+    border: 1px solid #2c6baa;
+    border-radius: 4px;
+    margin: 0 4px;  
+}
+  
+  .details-panel-body h2 {
+    margin-top: 0;
+    margin-bottom: 10px;
+    font-size: 24px;
+  }
+  
+  .details-panel-body p {
+    margin: 0;
+    font-size: 18px;
+    line-height: 1.5;
+  }
+  
+  .details-panel-poll {
+    margin-top: 20px;
+  }
+  
+  .details-panel-poll h3 {
+    margin-top: 0;
+    margin-bottom: 10px;
+    font-size: 18px;
+  }
+  
+  .details-panel-poll form label {
+    display: block;
+    margin-bottom: 10px;
+  }
+  
+  .details-panel-footer {
+    margin-top: 20px;
+  }
+  
+  .textbox {
+    background: #e5edf7
+  }
+
+  .fixed-button {
+    position: fixed;
+    top: 25px;
+    left: 25px;
+    color: #2c6baa;
+  }
+  
+  .hide {
+    display: none;
+  }
+  
+  
+
