+++ src/styles/votingPoll.css
@@ -1,0 +1,91 @@
+.voting-poll {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 20px;
+    border: 2px solid #ccc;
+    border-radius: 5px;
+    margin-bottom: 20px;
+  }
+  
+  .choices-container {
+    display: flex;
+    flex-direction: column;
+    margin-bottom: 20px;
+  }
+  
+  .choice {
+    display: flex;
+    align-items: center;
+    margin-bottom: 10px;
+  }
+  
+  label {
+    font-size: 16px;
+  }
+  
+  .vote-button {
+    background-color: #00ff51;
+    color: #fff;
+    border: none;
+    padding: 10px 20px;
+    border-radius: 5px;
+    cursor: pointer;
+    margin-bottom: 20px;
+    transition: all 0.2s ease-in-out;
+  }
+  
+  .vote-button:hover {
+    background-color: #0069d9;
+  }
+  
+  .results-container {
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .result {
+    display: flex;
+    align-items: center;
+    margin-bottom: 10px;
+  }
+  
+  .result-label {
+    font-size: 16px;
+    margin-right: 10px;
+  }
+  
+  .result-bar {
+    background-color: #007bff;
+    color: #fff;
+    height: 30px;
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+    padding-right: 10px;
+    border-top-right-radius: 5px;
+    border-bottom-right-radius: 5px;
+    transition: all 0.2s ease-in-out;
+  }
+  
+  .result-bar:hover {
+    background-color: #0069d9;
+  }
+  
+  .voting-poll-item-container {
+    background: #d0e1f2;
+    margin: 6px 0;
+    padding: 4px;
+    border-radius: 4px;
+    box-shadow: #2a58e2;
+    border: 1px solid black;
+  }
+  
+  .voting-poll-item {
+    display: flex;
+  }
+  
+  .voting-poll-item label {
+    width: 95%;
+    font-weight: bold;
+  }
